Rename PointCalculator, refactor tests
